+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # Changelog
+
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
@@ -7,10 +8,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ## [0.0.2] - 2021-09-07
+
 ### Added
 
 - EBS-03: tests for EBS-03
 - Github action to run coverage tests as PR
+- WEB3 type generation via hardhat
 
 ### Changed
 
@@ -26,8 +29,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - EBS-04: `kill` function has been removed
 
-
 ## [0.0.1] - 2021-08-26
+
 ### Added
 
 - Hardhat framework